pyproject.toml
[project]
name = "nanopng"
version = "0.1.0"
description = "Generate transparent PNG images from text prompts using Nano Banana Pro + background removal"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "google-genai",
    "Pillow",
    "rembg[cpu]",
    "python-dotenv",
]

[project.scripts]
nanopng = "nanopng.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nanopng"]

[dependency-groups]
dev = [
    "ruff",
    "pyright",
    "pytest",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.pyright]
typeCheckingMode = "strict"
include = ["src"]
exclude = ["**/node_modules", "**/__pycache__", ".venv"]
venvPath = "."
venv = ".venv"
